Hope it helps. You were so close!!
Problem was when you entered in this case, you're returning `getPins()` inside the map function, but not returning anything outside, so map was executed properly and so was getPins inside map, but once map execution finished, the `getPins()` called inside console.log returned nothing, so result was `undefined`. You could do `return arr.map(...` and will get result, but for the right solution you just want the last result of the map, that's why I reassign this result var and returns what's inside once all the iterations inside map ended. Also works with `forEach()`

If you don't see what was happening and why let me know and I'll try to clarify it to you